Fixed a bug related to "dirty" git repositories
The first version was built upon the assumption that the git repository
was freshly converted from and SVN one, so it only contains commits that
has a matching SVN revision. Running it on a "dirty" git repository,
which contains additional (non-SVN) commits results in an exeption as it
fails to find a revision ID in the commit message during the dictionary
building process. This commit fixes this situation by simply ignoring
those commits.